+++ www/Library/admin_widgets/ParagraphEditor/ParagraphEditor.php
@@ -8,13 +8,194 @@
  */
 namespace Library\admin_widgets\ParagraphEditor;
 
-use Library\views\Widget\Widget;
+use Library\views\Widget\Widget,
+    Boolive\values\Rule;
 
 class ParagraphEditor extends Widget
 {
+    public function getInputRule()
+    {
+        return Rule::arrays(array(
+                'REQUEST' => Rule::arrays(array(
+                        // Отображаемый объект или над которым выполняется действие
+                        'object' => Rule::entity(array('is', '/Library/content_samples/paragraphs/TextBlock'))->default($this->object)->required(),
+                        'call' => Rule::string(),
+                        // Аргументы вызываемых методов (call)
+                        'devide' => Rule::arrays(array(
+                            'value1' => Rule::string()->default('')->required(),
+                            'value2' => Rule::string()->default('')->required()
+                        )),
+                        'merge' => Rule::arrays(array(
+                            'primary' => Rule::entity(array('is', '/Library/content_samples/paragraphs/TextBlock'))->required(),
+                            'secondary' => Rule::entity(array('is', '/Library/content_samples/paragraphs/TextBlock'))->required()
+                        )),
+                        'save' => Rule::arrays(array(
+                            'value' => Rule::string()->default('')->required()
+                        )),
+                    )
+                )
+            )
+        );
+    }
+
     public function work($v = array())
     {
-        $v['value'] = $this->_input['REQUEST']['object']->value();
-        return parent::work($v);
+        if (isset($this->_input['REQUEST']['call'])){
+            // Отправка атрибутов
+            if ($this->_input['REQUEST']['call'] == 'load'){
+                 return array('attrib'=>$this->callLoad());
+            }else
+            // Разделение
+            if (isset($this->_input['REQUEST']['devide'])){
+                return $this->callDevide(
+                    $this->_input['REQUEST']['object'],
+                    $this->_input['REQUEST']['devide']['value1'],
+                    $this->_input['REQUEST']['devide']['value2']
+                );
+            }else
+            // Соединение
+            if (isset($this->_input['REQUEST']['merge'])){
+                 return $this->callMerge(
+                     $this->_input['REQUEST']['merge']['primary'],
+                     $this->_input['REQUEST']['merge']['secondary']
+                 );
+            }else
+            // Сохранение атрибутов
+            if (isset($this->_input['REQUEST']['save'])){
+                return $this->callSave(
+                    $this->_input['REQUEST']['object'],
+                    $this->_input['REQUEST']['save']['value']
+                );
+            }
+            return null;
+        }else{
+            $v['attrib'] = $this->callLoad();
+            $v['object'] = $this->_input['REQUEST']['object']->uri();
+            return parent::work($v);
+        }
+    }
+
+    /**
+     * Объединение
+     * @param \Boolive\data\Entity $primary
+     * @param \Boolive\data\Entity $secondary
+     * @return array
+     */
+    protected function callMerge($primary, $secondary)
+    {
+        $primary->value($primary->value().$secondary->value());
+        $primary->save();
+        $secondary->isDelete(true);
+        $secondary->save();
+        return true;
+    }
+
+    /**
+     * Раздление текстового блока на два
+     * @param \Boolive\data\Entity $object Объект разделяемого текстового блока
+     * @param $value1 Значение для первого блока
+     * @param $value2 Значение для второго блока
+     * @return array
+     */
+    protected function callDevide($object, $value1, $value2)
+    {
+        $obj = array();
+        $obj[0] = $object;
+        // Обновить (обрезать) значение объекта
+        $obj[0]->value($value1);
+        // Создать новый объект с прототипом как у $obj со вторым значением
+        // Отправить uri объектов или сгенерировать их html???
+        $obj[1] = $object->proto()->birth($object->parent());
+        $obj[1]->order($object->order()+1);
+        $obj[1]->value($value2);
+
+        $result = array();
+        foreach ($obj as $o){
+            $v = array();
+            $o->save();
+            $this->_input['REQUEST']['object'] = $o;
+            $v['attrib'] = $this->callLoad();
+            $v['object'] = $o->uri();
+            $result[] = array(
+                'html' => parent::work($v),
+                'uri' => $o->uri()
+            );
+        }
+        return $result;
+    }
+
+    /**
+     * Сохранение значения объекта
+     * @param \Boolive\data\Entity $object Сохраняемый объект
+     * @param string $value Новое значение
+     * @return mixed
+     */
+    protected function callSave($object, $value)
+    {
+        $v = array();
+        $object->value($value);
+        // Проверка и сохранение
+        /** @var $error \Boolive\errors\Error */
+        $error = null;
+        $object->save(false, false, $error);
+        if (isset($error) && $error->isExist()){
+            $v['error'] = array();
+            if ($error->isExist('_attribs')){
+                foreach ($error->_attribs as $key => $e){
+                    /** @var $e \Boolive\errors\Error */
+                    $v['error'][$key] = $e->getUserMessage(true,' ');
+                }
+                $error->delete('_attribs');
+            }
+            $v['error']['_other_'] = $error->getUserMessage(true);
+        }else{
+            $v['attrib'] = $this->callLoad();
+        }
+        return $v;
+    }
+
+    /**
+     * Отправка атрибутов объекта
+     * @return array
+     */
+    protected function callLoad()
+    {
+        /** @var $obj \Boolive\data\Entity */
+        $obj = $this->_input['REQUEST']['object'];
+        $v['tag'] = 'p';
+        if ($obj->is('/Library/content_samples/paragraphs/H1')){
+            $v['tag'] = 'h1';
+        }else
+        if ($obj->is('/Library/content_samples/paragraphs/H2')){
+            $v['tag'] = 'h2';
+        }else
+        if ($obj->is('/Library/content_samples/paragraphs/H3')){
+            $v['tag'] = 'h3';
+        }else
+        if ($obj->is('/Library/content_samples/paragraphs/H4')){
+            $v['tag'] = 'h4';
+        }else
+        if ($obj->is('/Library/content_samples/paragraphs/H5')){
+            $v['tag'] = 'h5';
+        }else
+        if ($obj->is('/Library/content_samples/paragraphs/H6')){
+            $v['tag'] = 'h6';
+        }else
+        if ($obj->is('/Library/content_samples/paragraphs/Head')){
+            $v['tag'] = 'h1';
+        }
+        if ($obj->is('/Library/content_samples/paragraphs/Blockquote')){
+            $v['tag'] = 'blockquote';
+        }else
+        if ($obj->is('/Library/content_samples/paragraphs/Code')){
+            $v['tag'] = 'pre';
+        }
+        $v['style'] = $obj->style->getStyle();
+        $v['value'] = $obj->value();
+        if (!trim($v['value'])){
+            $v['value'] = ' ';
+        }
+        $v['uri'] = $obj->uri();
+        return $v;
     }
 }