+++ www/Library/content_widgets/Head/Head.php
@@ -13,6 +13,24 @@ class Head extends Widget
     public function work($v = array())
     {
         $v['value'] = $this->_input['REQUEST']['object']->value();
+        $v['style'] = $this->_input['REQUEST']['object']->style->getStyle();
+        $obj = $this->_input['REQUEST']['object'];
+        $v['tag'] = 'h1';
+        if ($obj->is('/Library/content_samples/paragraphs/H2')){
+            $v['tag'] = 'h2';
+        }else
+        if ($obj->is('/Library/content_samples/paragraphs/H3')){
+            $v['tag'] = 'h3';
+        }else
+        if ($obj->is('/Library/content_samples/paragraphs/H4')){
+            $v['tag'] = 'h4';
+        }else
+        if ($obj->is('/Library/content_samples/paragraphs/H5')){
+            $v['tag'] = 'h5';
+        }else
+        if ($obj->is('/Library/content_samples/paragraphs/H6')){
+            $v['tag'] = 'h6';
+        }
         return parent::work($v);
     }
 }