+++ CHANGELOG.md
@@ -25,6 +25,7 @@ Version 2.0 work in progress
 - new #69: Added /Languages with English and Russian-language object (VladimirShestakov)
 - new #70: Added Session module (VladimirShestakov)
 - new: Added samples of paragraphs (P, H1-H2, Code, Blockquote) in /Library/content_samples/paragraphs (VladimirShestakov)
+- new #64: Added RichText editor. You can enter text, select, merge and split paragraphs (VladimirShestakov)
 - bug: Fixed getting proto and parent if they not exists. (VladimirShestakov)
 
 Version 2.0 preview October, 8 2012